+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Changed the `Media` closures `picture()` and `image()` to stack, like the cms `Artwork`'s
+
 ## 3.1.3 (September 25, 2026)
 
 - Fixed `AssetTrait::findAsset` for custom asset subclasses